Improve type safety for windowAPI
