+++ src-electron/windowAPI/index.ts
@@ -1,0 +1,10 @@
+export type WindowAPI = {
+  minimize: () => void;
+  toggleMaximize: () => void;
+  close: () => void;
+  pin: () => void;
+  unpin: () => void;
+  mute: () => void;
+  unmute: () => void;
+  openDevTools: () => void;
+};
